requirements.txt
streamlit==1.33.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.4
pandas==2.2.2
ollama==0.1.8
neo4j==5.17.0
reportlab==4.0.8
textwrap3==0.9.2
huggingface-hub==0.10.1
dotenv